Рефакторинг GZIP middleware: переименован пакет, улучшена обработка типов контента и добавлена поддержка сжатия ответов
